# backend/docs/swagger.yaml
definitions:
  models.ErrorMsg:
    properties:
      error:
        type: string
    type: object
  models.ExampleSentence:
    properties:
      chinese:
        type: string
      example:
        type: string
    type: object
...
  models.JapaneseWord:
    properties:
      chinese:
        type: string
      example:
        items:
          $ref: '#/definitions/models.ExampleSentence'
        type: array
      hiragana:
        type: string
      id:
        type: integer
      kanji:
        type: string
      katakana:
        type: string
      type:
        type: string
    type: object
...
  models.SearchResult-models_JapaneseWord:
    properties:
      count:
        type: integer
      page:
        type: integer
      pageSize:
        type: integer
      results:
        items:
          $ref: '#/definitions/models.JapaneseWord'
        type: array
    type: object
...
  models.SuccessMsg:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  title: Easy Japanese API
  version: "0.1"
paths:
...
  /api/words/{dictName}/accurate-search:
    post:
      consumes:
      - application/json
      description: Only ``kanji" and ``katakana" fields are used for comparison
      parameters:
      - description: Dictionary name
        in: path
        name: dictName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.JapaneseWord'
            type: array
        "400":
          description: Invalid JSON or Invalid dict name
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Check for similar words in the dictionary
      tags:
      - globalDictOp
  /api/words/{dictName}/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dictionary name
        in: path
        name: dictName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuccessMsg'
        "400":
          description: Invalid JSON or Invalid dict name
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "409":
          description: Duplicate word
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Insert word into dictionary
      tags:
      - globalDictOp
  /api/words/{dictName}/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dictionary name
        in: path
        name: dictName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMsg'
        "400":
          description: Invalid JSON or Invalid dict name
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Delete word in dictionary
      tags:
      - globalDictOp
  /api/words/{dictName}/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dictionary name
        in: path
        name: dictName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JapaneseWord'
        "400":
          description: Invalid JSON or Invalid dict name
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Update word in dictionary
      tags:
      - globalDictOp
  /api/words/{dictName}/fuzzy-search:
    get:
      consumes:
      - application/json
      parameters:
      - description: Dictionary name
        in: path
        name: dictName
        required: true
        type: string
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Results per page
        in: query
        name: RPP
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SearchResult-models_JapaneseWord'
        "400":
          description: Invalid JSON or Invalid dict name
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Check for similar words (fuzzy) in the dictionary
      tags:
      - globalDictOp
  /api/words/{dictName}/get:
    get:
      parameters:
      - description: Dictionary name
        in: path
        name: dictName
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Results per page
        in: query
        name: RPP
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SearchResult-models_JapaneseWord'
        "400":
          description: Invalid dictionary name
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Browse words in dictionary
      tags:
      - globalDictOp
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"